% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.statistics.R
\name{aggregate.statistics}
\alias{aggregate.statistics}
\title{This function splits a SingleCellExperiment along come conditions and computes summary statistics. This is useful when we want to look at the median expression of some markers on every cluster at a specific timepoint, for example. It almost entirely reuses code working in the background of `CATALYST::plotExprHeatmap()`}
\usage{
\method{aggregate}{statistics}(x, markers, by, clustering, fun = c("median", "mean", "sum"))
}
\arguments{
\item{x;}{a SingleCellDataset}

\item{markers;}{character vector: which markers do you wish to compute summary statistics for?}

\item{by;}{what condition to split the SingleCellExperiment by? has to be one or more of `colnames(colData(sce))`}

\item{clustering;}{which clustering should be used to split cells? has to be in of `names(cluster_codes(sce))`}

\item{fun;}{which summary statistc to calculate? has to be one of `"median"`, `"mean"` or `"sum"`}
}
\value{
a list of matrices (one for each cluster), where each column is a condition (e.g. timepoint), each row a cluster, and each value a summary statistic
}
\description{
This function splits a SingleCellExperiment along come conditions and computes summary statistics. This is useful when we want to look at the median expression of some markers on every cluster at a specific timepoint, for example. It almost entirely reuses code working in the background of `CATALYST::plotExprHeatmap()`
}
\examples{

ms <- aggregate_statistics(x=sce, markers=refined_markers, by = c("cluster_id", "timepoint"), fun="median", clustering="meta40")
}
\seealso{

}
