% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hard.downsample.R
\name{hard.downsample}
\alias{hard.downsample}
\title{This function performs hard downsampling of a flowset}
\usage{
hard.downsample(fs, event_number)
}
\arguments{
\item{fs;}{a flowCore flowset to be downsampled}

\item{event_number;}{a positive integer smaller than or equal to \code{min(fsApply(fs, nrow))}` indicating what number of cells to retain in the downsampled flowset}
}
\value{
a (smaller) flowCore flowset
}
\description{
It randomly selects a wiebke-defined number of cells from each fcs file in a flowset.
}
\note{
As the downsampling is performed randomly, running it multiple times will give slightly different results. To reproducibly have the same output include manually set the seed, as shown in the example below.
}
\examples{

# manually set the seed to have reproducible results; use whatever integer you like.
set.seed(1234)

# keep 5000 events from each file
smaller_vac69a <- hard.downsample(vac69a, 5000)

}
\seealso{
\code{\link{proportional.downsample}}
}
