% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportional.downsample.R
\name{proportional.downsample}
\alias{proportional.downsample}
\title{This function performs proportional downsampling of a flowset}
\usage{
proportional.downsample(fs, fraction)
}
\arguments{
\item{fs;}{a flowCore flowset to be downsampled}

\item{fraction;}{number between 0 and 1 indicating what fraction of cells to retain in the downsampled setwhat SingleCellExperiment was used for diffcyt?}
}
\value{
a flowCore flowset
}
\description{
i.e. it randomly selects a wiebke-defined percentage from each fcs file in a flowset.
}
\note{
As the downsampling is performed randomly, running it multiple times will give slightly different results. To reproducibly have the same output include manually set the seed, as shown in the example below.
}
\examples{

# manually set the seed to have reproducible results; use whatever integer you like.
set.seed(1234)

# keep half of all events
smaller_vac69a <- proportional.downsample(vac69a, 0.5)

}
\seealso{
\code{\link{hard.downsample}}
}
