% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.aggregate.R
\name{simplify.aggregate}
\alias{simplify.aggregate}
\title{This function splits a SingleCellExperiment along come conditions and computes summary statistics. This is useful when we want to look at the median expression of some markers on every cluster at a specific timepoint, for example. It almost entirely reuses code working in the background of `CATALYST::plotExprHeatmap()`}
\usage{
simplify.aggregate(ms, fun)
}
\arguments{
\item{x;}{a list of matrices made with aggregate.statistics}

\item{fun;}{which summary statistc to calculate across clusters? has to be one of `"median"`, `"mean"` or `"sum"`}
}
\value{
a single matrix, where each column is a condition (e.g. timepoint) with an additional column for the summary statistic, each row a cluster, and each value a summary statistic calculated with `aggregate.statistics` or `simplify.aggregate`.
}
\description{
This function splits a SingleCellExperiment along come conditions and computes summary statistics. This is useful when we want to look at the median expression of some markers on every cluster at a specific timepoint, for example. It almost entirely reuses code working in the background of `CATALYST::plotExprHeatmap()`
}
\examples{

median_ms <- simplify.aggregate(ms, "median")
}
\seealso{
\code{\link{aggregate.statistics}}
}
