% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vac69a_single_cell_experiment}
\alias{vac69a_single_cell_experiment}
\title{wiebkeTOF practise data}
\format{
A SingleCellExperiment
\describe{
  \item{fcs_files}{24 files of 5000 cells each (120000 cells in total)}
  \item{metadata}{a dataframe associating file names with volunteer code, timepoint, batch}
  \item{panel}{a dataframe with channel and marker names}
}
}
\source{
<link will come>
}
\usage{
vac69a_single_cell_experiment
}
\description{
5000 cells each from baseline, C10, DoD & T6 samples from the six volunteers (v02, v03, v05, v06, v07, v09) of vac69a
}
\keyword{datasets}
